fix: gitignore, docker-compose, CI fix(api): image, feat: alembic

+++ app/schemas/product.py
@@ -2,7 +2,7 @@ from pydantic import BaseModel
 
 class Product(BaseModel):
     name: str
-    calories: int
-    proteins: int
-    fats: int
-    carbs: int
+    calories: float
+    proteins: float
+    fats: float
+    carbs: float
